development

+++ src/main/java/com/app/controller/UserController.java
@@ -1,23 +1,19 @@
 package com.app.controller;
 
-import com.app.entity.QUser;
 import com.app.entity.Role;
 import com.app.entity.User;
 import com.app.repository.RoleRepository;
 import com.app.repository.UserRepository;
+import com.app.searchform.SearchStrategy;
 import com.app.service.UserService;
+import com.app.utils.PaginationDetails;
 import com.app.utils.Result;
-import com.querydsl.core.BooleanBuilder;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.util.Strings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.StringTrimmerEditor;
 import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -29,11 +25,8 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
-import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 
 
@@ -48,111 +41,63 @@ public class UserController {
 	
 	@Autowired
 	RoleRepository roleRepository;
-	
-	//@Autowired
-	//ApplicationEventPublisher applicationEventPublisher;
-	
-	//@Autowired
-	//private EmailService emailService;
-	
-	//private BCryptPasswordEncoder bCryptPasswordEncoder;
-	
-	@Autowired
-	private UserService userService;
-	
-	private final int[] PAGE_SIZES = {5,10,20};
-	
-	@RequestMapping(value="registration/{returnPage}", method = RequestMethod.GET)
-	public String register(Model model, @PathVariable(name = "returnPage", required = false)  Optional<String> returnPage) {
-		model.addAttribute("user", new User());	
-		
-		if(returnPage.isPresent())
-			model.addAttribute("returnPage", returnPage.get());
-		
-		return "user/userRegistration";
-	}
-	
-	@InitBinder
-	public void initBinder ( WebDataBinder binder )
-	{
-	    StringTrimmerEditor stringtrimmer = new StringTrimmerEditor(true);  
-	    binder.registerCustomEditor(String.class, stringtrimmer);
-	}
-	
-	
-	@RequestMapping(value="/list")
-	public String list(@RequestParam(name = "page", defaultValue = "1", required = false) int page,
-			@RequestParam(name = "size", required = false, defaultValue = "10") int size,
-			@RequestParam(name = "orderBy", required = false, defaultValue = "id") String orderBy,
-			@RequestParam(name = "sort", required = false, defaultValue = "ASC") String sort,
-			@RequestParam(name = "searchArguments",required = false, defaultValue = "&") String searchArguments, 
-			@ModelAttribute("user")  User user,
-			Model model) {
-		
-		user.prepareFiledsForSearch();
-		
-		BooleanBuilder queryBuilder = new BooleanBuilder();
-		searchArguments = prepareQueryAndSearchArguments(user, searchArguments, queryBuilder);
-		
-		PageRequest pageable = null;
-		pageable = PageRequest.of(page, size, Direction.fromString(sort), orderBy);
-	
-		Page<User> userPage = userRepository.findAll(queryBuilder, pageable);
-		
-        if(userPage.getTotalPages() > 0) {
-            List<Integer> pageNumbers = IntStream.rangeClosed(1 ,userPage.getTotalPages()).boxed().collect(Collectors.toList());
-            model.addAttribute("numbers", pageNumbers);
-        }
-        
-        
-        model.addAttribute("pages", userPage);
-        model.addAttribute("orderBy", orderBy);
-        model.addAttribute("sort", sort);
-        model.addAttribute("searchArguments", searchArguments);
-        model.addAttribute("page", page);
-        model.addAttribute("size", size);
-        model.addAttribute("pageSizes", PAGE_SIZES);
-        model.addAttribute("user", user);
-		
-		return "/user/userList";
-	}
-	
-	private String prepareQueryAndSearchArguments(User user, String searchArguments, BooleanBuilder queryBuilder) {
-		StringBuilder stringBuilder = new StringBuilder(searchArguments);
-		
-		if(StringUtils.isNotBlank(user.getLogin())) {
-			queryBuilder.and(QUser.user.login.contains(user.getLogin()));
-			stringBuilder.append("login=").append(user.getLogin()).append("&");
-		}
-		if(StringUtils.isNotBlank(user.getEmail())) {
-			queryBuilder.and(QUser.user.email.contains(user.getEmail()));
-			stringBuilder.append("email=").append(user.getEmail()).append("&");
-		}
-		if(StringUtils.isNotBlank(user.getFirstName())) {
-			queryBuilder.and(QUser.user.firstName.contains(user.getFirstName()));
-			stringBuilder.append("firstName=").append(user.getFirstName()).append("&");
-		}
-		if(StringUtils.isNotBlank(user.getLastName())) {
-			queryBuilder.and(QUser.user.lastName.contains(user.getLastName()));
-			stringBuilder.append("lastName=").append(user.getLastName()).append("&");
-		}
-		if(user.getActive() != null) {
-			queryBuilder.and(QUser.user.active.eq(user.getActive()));
-			stringBuilder.append("active=").append(user.getActive()).append("&");
-		}
-		if(StringUtils.isNotBlank(user.getCity())) {
-			queryBuilder.and(QUser.user.city.contains(user.getCity()));
-			stringBuilder.append("city=").append(user.getCity()).append("&");
-		}
-		
-		return stringBuilder.toString();
-	}
-	
-	
+
+    //@Autowired
+    //ApplicationEventPublisher applicationEventPublisher;
+
+    //@Autowired
+    //private EmailService emailService;
+
+    //private BCryptPasswordEncoder bCryptPasswordEncoder;
+
+    @Autowired
+    private UserService userService;
+
+    private final SearchStrategy<User> userSearchStrategy;
+
+    private final int[] PAGE_SIZES = {5, 10, 20};
+
+    public UserController(SearchStrategy<User> userSearchStrategy) {
+        this.userSearchStrategy = userSearchStrategy;
+    }
+
+    @RequestMapping(value = "registration/{returnPage}", method = RequestMethod.GET)
+    public String register(Model model, @PathVariable(name = "returnPage", required = false) String returnPage) {
+        model.addAttribute("user", new User());
+
+        if (returnPage != null)
+            model.addAttribute("returnPage", returnPage);
+
+
+        return "user/userRegistration";
+    }
+
+    @InitBinder
+    public void initBinder(WebDataBinder binder) {
+        StringTrimmerEditor stringtrimmer = new StringTrimmerEditor(true);
+        binder.registerCustomEditor(String.class, stringtrimmer);
+    }
+
+
+    @RequestMapping(value = "/list")
+    public String list(@RequestParam(name = "page", defaultValue = "1", required = false) int page,
+                       @RequestParam(name = "size", required = false, defaultValue = "10") int size,
+                       @RequestParam(name = "orderBy", required = false, defaultValue = "id") String orderBy,
+                       @RequestParam(name = "sort", required = false, defaultValue = "ASC") String sort,
+                       @RequestParam(name = "searchArguments", required = false, defaultValue = "&") String usedSearchArguments,
+                       @ModelAttribute("user") User user,
+                       Model model) {
+
+        model.addAttribute("requestMapping", "list");
+
+        PaginationDetails paginationDetails = PaginationDetails.builder().page(page).size(size).orderBy(orderBy).sort(sort).build();
+        model.addAllAttributes(userSearchStrategy.prepareSearchForm(user, paginationDetails));
+
+        return "/user/userList";
+    }
 	
 	@RequestMapping(value="edit/{id}", method = RequestMethod.GET)
     public String edit(@PathVariable("id") Long id, Model model) {
-	//public String edit(@PathVariable("id") Long id, BindingResult bindingResult, Model model, Errors errors) throws Exception { 
 		Optional<User> user = userRepository.findById(id);
 		
 		if(user.isPresent()) 
@@ -211,27 +156,26 @@ public class UserController {
 			
 			return "redirect:/user/registrationSuccess";
 		}
-		else {
-		//	result.getValidationResult().entrySet().stream().forEach(e -> bindingResult.reject(e.getKey(), e.getValue()));
-		//	result.getValidationResult().entrySet().stream().peek(e -> System.out.println("error, key " + e.getKey() + " value " + e.getValue()));
-			
-		//	if(errors.getAllErrors() != null)
-		//		errors.getAllErrors().addAll(result.getValidationResult().entrySet().stream().map(e ->  new ObjectError(e.getKey(), e.getValue())).collect(Collectors.toList()));
+//		else {
+        //	result.getValidationResult().entrySet().stream().forEach(e -> bindingResult.reject(e.getKey(), e.getValue()));
+        //	result.getValidationResult().entrySet().stream().peek(e -> System.out.println("error, key " + e.getKey() + " value " + e.getValue()));
+
+        //	if(errors.getAllErrors() != null)
+        //		errors.getAllErrors().addAll(result.getValidationResult().entrySet().stream().map(e ->  new ObjectError(e.getKey(), e.getValue())).collect(Collectors.toList()));
 
 
+        //	for (Entry<String, ValidationDetails> entry : result.getValidationResult().entrySet()) {
 
-		//	for (Entry<String, ValidationDetails> entry : result.getValidationResult().entrySet()) {
+        // o to co trzeba poprawic po zmiane na ResultDetails  przestalo dzialac
+        //bindingResult.addError(new FieldError("user", entry.getKey(), "", false, null, null, entry.getValue().name()));
+        //		}
 
-				// o to co trzeba poprawic po zmiane na ResultDetails  przestalo dzialac
-				//bindingResult.addError(new FieldError("user", entry.getKey(), "", false, null, null, entry.getValue().name()));
-			}
-			
-			model.addAttribute("user", user);
-			model.addAttribute("error", bindingResult);
-			
-			return "user/registerUser";
+        model.addAttribute("user", user);
+        model.addAttribute("error", bindingResult);
 
-	}
+        return "user/registerUser";
+
+    }
 	
 	@RequestMapping(value="adminSettings/{id}", method = RequestMethod.GET)
 	public String adminSettings(@PathVariable("id") Long id, Model model) {
@@ -249,68 +193,35 @@ public class UserController {
 	
 	@RequestMapping(value="delete/{id}", method = RequestMethod.GET)
 	public String delete(@PathVariable("id") Long id, Model model, RedirectAttributes redirectAttributes) {
-		//public String edit(@PathVariable("id") Long id, BindingResult bindingResult, Model model, Errors errors) throws Exception { 
-		Optional<User> user = userRepository.findById(id);
-		
-		if(user.get().getAnnouncementList().size() > 0) {
-			redirectAttributes.addFlashAttribute("message", "user ma przypisane aukcje, zanim usuniesz usuera trzeba je skasowac");
-			
-			return "redirect:/user/edit/" + user.get().getId();
-		}
-		else {
-			redirectAttributes.addFlashAttribute("message", "konto usuniete");
-		 	userRepository.delete(user.get());
-			
-			return "redirect:/user/list";
-		}
+        //public String edit(@PathVariable("id") Long id, BindingResult bindingResult, Model model, Errors errors) throws Exception {
+        Optional<User> user = userRepository.findById(id);
+
+
+        if (user.isPresent() && user.get().getAnnouncementList().size() > 0) {
+            redirectAttributes.addFlashAttribute("message", "user ma przypisane aukcje, zanim usuniesz usuera trzeba je skasowac");
+
+            return "redirect:/user/edit/" + user.get().getId();
+        } else {
+            redirectAttributes.addFlashAttribute("message", "konto usuniete");
+            userRepository.delete(user.get());
+
+            return "redirect:/user/list";
+        }
 	}
-	
-/*	private Model createUser(User user, Model model, BindingResult bindingResult) {
-		if (bindingResult.hasErrors()) {
-			for (ObjectError objectError : bindingResult.getAllErrors()) {
-				logger.error(objectError.toString());
-			}
-		} 
-		else if(!user.getPassword().equals(user.getPasswordConfirm())) {
-			model.addAttribute("passwordsAreNotSame",true);
-		}
-		else if (userRepository.findByLogin(user.getLogin()) != null) {
-			model.addAttribute("loginAlreadyExists", true);
-		}
-		else if (userRepository.countByEmail(user.getEmail()) > 0) {
-			model.addAttribute("emailAlreadyExists", true);
-		}
-	    else if (!bindingResult.hasErrors()) {
-	    	Role userRole = roleRepository.findByName("ROLE_USER");
-	    	
-	    	if(userRole == null)
-	    		throw new Exception("Role w systemie nie sa zdefiniowane");
-	    	
-	    	user.setRole(userRole);
-	    	user.setActive(true);
-	    	user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
-	    	userRepository.save(user);
-	    	
-	    	model.addAttribute("user", user);
-	    }
-		
-	    	return model;
-	}*/
 	
 	@RequestMapping(value="update", method = RequestMethod.POST)
 	public String update(@ModelAttribute("user") @Validated({User.ValidateAllFieldsWithoutPass.class}) User user, BindingResult bindingResult, Model model, Errors errors) {
-		
-		if(errors.hasErrors() || !validateEmailBeforeUpdate(user, model)) {
-			model.addAttribute("user", user);
-			model.addAttribute("error",errors);
-			return "user/userEdit";
-		}
-		else {
-			Optional<User> userFromDB = userRepository.findById(user.getId());
-			
-			if(userFromDB.isPresent()) {
-				user.setPassword(userFromDB.get().getPassword());
-				userRepository.save(user);
+
+        if (errors.hasErrors() || validateEmailBeforeUpdate(user, model)) {
+            model.addAttribute("user", user);
+            model.addAttribute("error", errors);
+            return "user/userEdit";
+        } else {
+            Optional<User> userFromDB = userRepository.findById(user.getId());
+
+            if (userFromDB.isPresent()) {
+                user.setPassword(userFromDB.get().getPassword());
+                userRepository.save(user);
 			}
 			else
 				throw new NoSuchElementException ();
@@ -321,18 +232,17 @@ public class UserController {
 	
 	@RequestMapping(value="changePass", method = RequestMethod.POST)
 	public String changePass(@ModelAttribute("user") @Validated({User.ValidatePassOnly.class}) User user, BindingResult bindingResult, Model model, Errors errors) {
-		
-		if(errors.hasErrors() || !validateEmailBeforeUpdate(user, model)) {
-			model.addAttribute("user", user);
-			model.addAttribute("error",errors);
-			return "user/userChangePass";
-		}
-		else {
-			Optional<User> userFromDB = userRepository.findById(user.getId());
-			
-			if(userFromDB.isPresent()) {
-		//		userFromDB.get().setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
-				// do zmiany
+
+        if (errors.hasErrors() || validateEmailBeforeUpdate(user, model)) {
+            model.addAttribute("user", user);
+            model.addAttribute("error", errors);
+            return "user/userChangePass";
+        } else {
+            Optional<User> userFromDB = userRepository.findById(user.getId());
+
+            if (userFromDB.isPresent()) {
+                //		userFromDB.get().setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+                // do zmiany
 				userRepository.save(userFromDB.get());
 			}
 			else
@@ -355,12 +265,12 @@ public class UserController {
 	}
 	
 	private boolean validateEmailBeforeUpdate(User user, Model model) {
-		if(userRepository.countByEmailAndIdNot(user.getEmail(), user.getId()) > 0) {
-			model.addAttribute("emailAlreadyExists", true);
-			return false;
-		}
-			return true;
-	}
+        if (userRepository.countByEmailAndIdNot(user.getEmail(), user.getId()) > 0) {
+            model.addAttribute("emailAlreadyExists", true);
+            return true;
+        }
+        return false;
+    }
 	
 	
 	// @Validated(User.ValidateAllFieldsWithoutPass.class)
